requirements.txt
wandb
ffcv
torchsummary